feat: Refactor bot status command and add custom status feature
- Renamed `setstatus.js` to `set.js` for better command organization.
- Changed the `status` subcommand to `activity` for managing bot activities.
- Introduced a new `status` subcommand allowing admins to set a custom status.
- Added duration options for both subcommands, allowing temporary or permanent status updates.
- Improved logging system to track status and activity changes.
